requirements.txt
absl-py==1.4.0
appnope==0.1.3
asttokens==2.2.1
autograd==1.6.2
backcall==0.2.0
cma==2.7.0
comm==0.1.4
contourpy==1.1.0
cycler==0.11.0
debugpy==1.6.7
decorator==5.1.1
executing==1.2.0
fonttools==4.41.1
future==0.18.3
ijson==3.2.3
importlib-metadata==6.8.0
importlib-resources==6.0.0
ipykernel==6.25.0
ipython==8.14.0
jedi==0.19.0
jupyter_client==8.3.0
jupyter_core==5.3.1
kiwisolver==1.4.4
matplotlib==3.7.2
matplotlib-inline==0.1.6
nest-asyncio==1.5.7
numpy==1.25.1
ortools==9.7.2996
packaging==23.1
pandas==2.0.3
parso==0.8.3
pexpect==4.8.0
pickle5==0.0.11
pickleshare==0.7.5
Pillow==10.0.0
platformdirs==3.10.0
prompt-toolkit==3.0.39
protobuf==4.24.2
psutil==5.9.5
psycopg2-binary==2.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.15.1
pymoo==0.5.0
pyparsing==3.0.9
Pyro5==5.14
python-dateutil==2.8.2
pytz==2023.3
pyzmq==25.1.0
scipy==1.11.1
serpent==1.41
six==1.16.0
stack-data==0.6.2
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
typing_extensions==4.7.1
tzdata==2023.3
wcwidth==0.2.6
zipp==3.16.2
